cleanup, add new script bcs.py
